backend/requirements.txt
fastapi
uvicorn
numpy
requests
web3
matplotlib
pydantic[email]
eth-utils
# Auth + DB
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
